+++ crates/wasmtime/src/runtime/vm/vmcontext.rs
@@ -333,7 +333,7 @@ pub struct VMTableDefinition {
     pub base: *mut u8,
 
     /// The current number of elements in the table.
-    pub current_elements: u32,
+    pub current_elements: usize,
 }
 
 #[cfg(test)]
